refactor: repair solid maintenance hotspots
