--- src/components/naver-map/naverMapHelpers.ts
@@ -1,11 +1,0 @@
-export const DAEJEON_CENTER = { latitude: 36.3504, longitude: 127.3845 };
-
-export { loadNaverMaps } from './mapSdk';
-export {
-  currentLocationMarkerContent,
-  festivalMarkerContent,
-  hasFestivalCoordinates,
-  placeMarkerContent,
-  routeStepMarkerContent,
-} from './markerContent';
-export { getSelectionVerticalOffset } from './selectionOffset';
